Add back no-repeat for arrow keys
Accidentally removed no-repeat functionality for navigation across panes
with arrow keys. This is particularly annoying in the scenario where we
move up to a pane above by pressing Prefix-Up, and then immediately
pressing Up to run the previous command. When the repeat functionality
is switched on, the second Up doesn't cycle the previous command, but
instead goes up another pane, i.e. tmux interprets the series of
commands as Prefix-Up-Up instead of Prefix-Up Up. Soluton is to disable
repeat function.