Manually write snippet into .vimrc
Previously there was a snippet for invoking the Python debugger pdb,
which was stored in a text file. Now we shift this short line of code
into the .vimrc directly. Not pretty, but it's the only solution to keep
things aligned properly when invoking this pdb snippet.

We'll see how things go, before deciding which method is best. That is,
to use the old method of storing the snippet somewhere else in a text
file, or just hard-code into .vimrc directly.