Add .gitignore
Because the config files like .vimrc, etc are at the root directory, git
status lists out all the untracked files, which can be quite annoying.
It's also quite inconvenient having to pass in the -uno flag all the
time. So we use this .gitignore to ignore all files, except the config
files and this .gitignore we want to track.