Change checkout alias
Previously, alias for checkout was 'co', which was very similar to the
alias for commit 'ci'. Because the keys 'i' and 'o' are next to each
other, it is very easy to accidentally type 'co' when what we really
want is 'ci'. To fix this annoying behaviour, we change the alias of
checkout from 'co' to 'cho'.